add reservation form, buttons, media objects and labels
Updated the navigation bar to include a form to allow the user to type
in their login information to be able to log into the website.

Added a button to the Jumbotron to enable users to reserve a table at
the restaurant. Clicking on this button should forward to the
reservation form at the bottom of the page.

Designed a form to enable users to submit a reservation request for a
table. The row should also have included a dismissible warning alert.
Formatted the contents of the second row of the page using media class.
The content column of the row should have been converted to a media
object.